+++ monkey/infection_monkey/post_breach/actions/communicate_as_new_user.py
@@ -4,7 +4,8 @@ import random
 import string
 import subprocess
 
-from infection_monkey.utils.auto_new_user import NewUserError, create_auto_new_user
+from infection_monkey.utils.new_user_error import NewUserError
+from infection_monkey.utils.auto_new_user_factory import create_auto_new_user
 from common.data.post_breach_consts import POST_BREACH_COMMUNICATE_AS_NEW_USER
 from infection_monkey.post_breach.pba import PBA
 from infection_monkey.telemetry.post_breach_telem import PostBreachTelem
