+++ monkey/infection_monkey/utils/new_user_error.py
@@ -1,0 +1,2 @@
+class NewUserError(Exception):
+    pass
